Fix Connection Pool NoMethodError
This error would pop up seemingly randomly when open connections would become stale. The gem builds out a set of the connections that are available to rehydrate the session with, but because of the `include?` it would have the opportunity to compare a set of connections against a connection. This flattens the set so that we are comparing apples to apples